9p: Hide implementation details
